KakaoNavi를 이용한 이동 거리와 이동 시간 자동 계산 및 Dijkstra 알고리즘을 활용한 최적 허브 간 이동 경로 찾기 구현

+++ hub/src/main/java/com/team12/hub/hubPath/dto/HubPathResponseDto.java
@@ -16,7 +16,7 @@ public class HubPathResponseDto {
     private UUID id;
     private UUID fromHubId;  // 출발 허브 ID
     private UUID toHubId;  // 도착 허브 ID
-    private double distance; // 허브 간 거리
+    private int distance; // 허브 간 거리
     private int duration;  // 소요시간
 
     public HubPathResponseDto(HubPath hubPath){
